Replace two remaining custom libdefs
meta:sha:cd5b02f85d520d71dfceaab1ce0a738f43316ffc
